refactor: move compare_epoch_id to EpochManagerAdapter
cc #6910
